Chatbot/requirements.txt
streamlit
requests
flask
flask-cors
pymongo
bson
python-dotenv
langchain
google-generativeai
